+++ Source/Core/Extent.js
@@ -354,22 +354,9 @@ define([
         return this.west >= this.east || this.south >= this.north;
     };
 
-    /**
-     * Samples this Extent so that it includes a list of Cartesian points suitable for passing to
-     * {@link BoundingSphere#fromPoints}.  Sampling is necessary to account
-     * for extents that cover the poles or cross the equator.
-     *
-     * @param {Ellipsoid} [ellipsoid=Ellipsoid.WGS84] The ellipsoid to use.
-     * @param {Array} [result] The array of Cartesians onto which to store the result.
-     * @return {Array} The modified result parameter or a new Array of Cartesians instances if none was provided.
-     */
-    Extent.prototype.subsample = function(ellipsoid, result) {
-        return this.subsampleWithHeight(ellipsoid, 0.0, result);
-    };
-
     var subsampleLlaScratch = new Cartographic();
     /**
-     * Samples this Extent at a height above the ellipsoid so that it includes a list of Cartesian points suitable for passing to
+     * Samples this extent so that it includes a list of Cartesian points suitable for passing to
      * {@link BoundingSphere#fromPoints}.  Sampling is necessary to account
      * for extents that cover the poles or cross the equator.
      *
@@ -378,7 +365,7 @@ define([
      * @param {Array} [result] The array of Cartesians onto which to store the result.
      * @return {Array} The modified result parameter or a new Array of Cartesians instances if none was provided.
      */
-    Extent.prototype.subsampleWithHeight = function(ellipsoid, surfaceHeight, result) {
+    Extent.prototype.subsample = function(ellipsoid, surfaceHeight, result) {
         ellipsoid = defaultValue(ellipsoid, Ellipsoid.WGS84);
         surfaceHeight = defaultValue(surfaceHeight, 0.0);