+++ Source/Core/ExtentGeometry.js
@@ -317,7 +317,7 @@ define([
          *
          * @type BoundingSphere
          */
-        this.boundingSphere = BoundingSphere.fromExtentWithHeight3D(extent, ellipsoid, surfaceHeight);
+        this.boundingSphere = BoundingSphere.fromExtent3D(extent, ellipsoid, surfaceHeight);
 
         /**
          * The 4x4 transformation matrix that transforms the geometry from model to world coordinates.